refactor(HIGH-28): mergeUniqueContents 함수 utils로 분리
